config: change to legacy handler
